cleanup(intents) clean top intents
